fixed dtype in dcn_v2

+++ layers/dcn_v2.py
@@ -102,24 +102,24 @@ class DCNv2(tf.keras.layers.Layer):
         oyox, mask = offset[..., :2*self.ks], offset[..., 2*self.ks:]
         mask = tf.nn.sigmoid(mask)
         #[H, W, 2]
-        grid_yx = tf.stack(tf.meshgrid(tf.range(iw), tf.range(ih))[::-1], axis = -1)
+        grid_yx = tf.stack(tf.meshgrid(tf.range(iw), tf.range(ih))[::-1], axis=-1)
         #[1, H, W, 9, 2]
         grid_yx = tf.reshape(grid_yx, [1, ih, iw, 1, 2]) + self.phw + self.patch_yx
         #[B, H, W, 9, 2]
-        grid_yx = tf.cast(grid_yx, 'float32') + tf.reshape(oyox, [bs, ih, iw, -1, 2])
+        grid_yx = tf.cast(grid_yx, oyox.dtype) + tf.reshape(oyox, [bs, ih, iw, -1, 2])
         grid_iy0ix0 = tf.floor(grid_yx)
-        grid_iy1ix1 = tf.clip_by_value(grid_iy0ix0 + 1, 0, tf.constant([ih+1, iw+1], dtype = 'float32'))
+        grid_iy1ix1 = tf.clip_by_value(grid_iy0ix0 + 1, 0, tf.constant([ih+1, iw+1], dtype=grid_iy0ix0.dtype))
         #[B, H, W, 9, 1] * 2
         grid_iy1, grid_ix1 = tf.split(grid_iy1ix1, 2, axis = 4)
-        grid_iy0ix0 = tf.clip_by_value(grid_iy0ix0, 0, tf.constant([ih+1, iw+1], dtype = 'float32'))
+        grid_iy0ix0 = tf.clip_by_value(grid_iy0ix0, 0, tf.constant([ih+1, iw+1], dtype=grid_iy0ix0.dtype))
         grid_iy0, grid_ix0 = tf.split(grid_iy0ix0, 2, axis = 4)
-        grid_yx = tf.clip_by_value(grid_yx, 0, tf.constant([ih+1, iw+1], dtype = 'float32'))
+        grid_yx = tf.clip_by_value(grid_yx, 0, tf.constant([ih+1, iw+1], dtype=grid_yx.dtype))
         #[B, H, W, 9, 4, 1]
         batch_index = tf.tile(tf.reshape(tf.range(bs), [bs, 1, 1, 1, 1, 1]), [1, ih, iw, self.ks, 4, 1])
         #[B, H, W, 9, 4, 2]
         grid = tf.reshape(tf.concat([grid_iy1ix1, grid_iy1, grid_ix0, grid_iy0, grid_ix1, grid_iy0ix0], axis = -1), [bs, ih, iw, self.ks, 4, 2])
         #[B, H, W, 9, 4, 3]
-        grid = tf.concat([batch_index, tf.cast(grid, 'int32')], axis = -1)
+        grid = tf.concat([batch_index, tf.cast(grid, tf.int32)], axis = -1)
         #[B, H, W, 9, 2, 2]
         delta = tf.reshape(tf.concat([grid_yx - grid_iy0ix0, grid_iy1ix1 - grid_yx], axis = -1), [bs, ih, iw, self.ks, 2, 2])
         #[B, H, W, 9, 2, 1] * [B, H, W, 9, 1, 2] = [B, H, W, 9, 2, 2]
